Extracted gallery component :rocket:
