+++ src/pages/art.tsx
@@ -1,72 +1,17 @@
 import * as React from "react";
-import { useState } from "react";
 import { graphql, PageProps } from "gatsby";
 import { ArtQuery } from "../../graphql-types";
 import { PageWithContent } from "../components/design-system/templates/page-with-content";
 import { getCurrentLocationFrom } from "../logic/current-location";
 import { tracking } from "../logic/tracking";
 import { OgPageType } from "../logic/seo";
-import styled from "styled-components";
 import { PageTitle } from "../components/design-system/molecules/page-title";
-import { ContainerFluid } from "../components/design-system/atoms/container-fluid";
-import { GatsbyImage, getSrc, IGatsbyImageData } from "gatsby-plugin-image";
-import { Paragraph } from "../components/design-system/atoms/paragraph";
-import { artDescriptions } from "../logic/art";
-import { ModalWithImage } from "../components/design-system/organism/modal-with-image";
-
-const GalleryContainer = styled(ContainerFluid)`
-  padding: 0;
-  margin: 0 0 ${(props) => props.theme.spacing[7]};
-  display: grid;
-  align-items: center;
-  justify-items: center;
-  grid-column-gap: 20px;
-  grid-row-gap: 20px;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-
-  @media (min-width: 992px) {
-    grid-column-gap: 20px;
-    grid-row-gap: 40px;
-  }
-`;
-
-const GalleryImageFrame = styled.figure`
-  padding: ${(props) => props.theme.spacing[2]};
-  margin: 0;
-  background-color: ${(props) => props.theme.light.generalBackgroundLight};
-
-  @media (prefers-color-scheme: dark) {
-    background: ${(props) => props.theme.dark.generalBackgroundLight};
-  }
-`;
-
-const GalleryImageDescription = styled(Paragraph)`
-  width: 250px;
-  height: 55px;
-  text-align: center;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const GalleryImage = styled(GatsbyImage)`
-  height: auto;
-  width: 250px;
-  object-fit: cover;
-  transition: opacity 0.25s ease-in-out;
-
-  &:hover {
-    opacity: 0.7;
-  }
-`;
+import { Gallery } from "../components/design-system/organism/gallery";
 
 const Art: React.FC<PageProps<ArtQuery>> = ({ data, location }) => {
   const siteMetadata = data.site!.siteMetadata!;
   const author = siteMetadata.author!;
   const featuredImage = siteMetadata.featuredImage!;
-  const [currentImage, setCurrentImage] = useState<IGatsbyImageData | null>(
-    null
-  );
 
   return (
     <PageWithContent
@@ -79,28 +24,7 @@ const Art: React.FC<PageProps<ArtQuery>> = ({ data, location }) => {
       description={"Fabrizio Duroni art gallery"}
     >
       <PageTitle>My drawings</PageTitle>
-      <GalleryContainer>
-        {data.allFile.edges.map((image) => (
-          <GalleryImageFrame key={image.node.name}>
-            <GalleryImage
-              alt={image.node.name}
-              image={image.node.childImageSharp!.gatsbyImageData!}
-              onClick={() =>
-                setCurrentImage(image.node.childImageSharp!.gatsbyImageData!)
-              }
-            />
-            <GalleryImageDescription>
-              {artDescriptions[image.node.name]}
-            </GalleryImageDescription>
-          </GalleryImageFrame>
-        ))}
-      </GalleryContainer>
-      {currentImage && (
-        <ModalWithImage
-          imageUrl={getSrc(currentImage)!}
-          onClick={() => setCurrentImage(null)}
-        />
-      )}
+      <Gallery images={data.allFile.edges} />
     </PageWithContent>
   );
 };
